Extract tests for `@JSExportTopLevel` in a separate file/class.
`@JSExportTopLevel` originally evolved from `@JSExport`. At the
time, it made sense to have their tests in the same test class.
Nowadays, however, these features have more differences than things
in common. These differences will be even more exacerbated with
WebAssembly, since it will support `@JSExportTopLevel` but not
`@JSExport`.
